Factor out apache auth checking.
	modified:   src/python/CertWebServer.py
	new file:   src/python/apache_utils.py
	modified:   src/python/services/RequestsDB.py
